Do initial refactor of data dictionary service
In order to fix issues for #706, additional parameters will need to be
implemented in the REST api. These will come with their own changes that
may be significant.

Perform an initial refactor of the data dictionary service with the
following goals:

- Reduce method signature size between DataDictionaryOperations and
DataDictionary with ConnectionConfig class.
- Extract DefaultMetadataField transformation to separate class to
improve code clarity and make testing easier.
- Reduce duplicated code where possible.
- Add more unit tests and documentation.